clientlib:split main.js into {anonvote, census}.js
