+++ clientlib/package.json
@@ -2,7 +2,6 @@
 	"name": "anon-vote-clientlib",
 	"version": "0.0.1",
 	"description": "anon-vote clientlib",
-	"main": "./src/main.js",
 	"directories": {
 		"test": "test"
 	},
